Improved Output Model, by exposing more information and operations

+++ server/src/main/kotlin/pt/isel/g20/unicommunity/template/service/TemplateService.kt
@@ -3,6 +3,7 @@ package pt.isel.g20.unicommunity.template.service
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import pt.isel.g20.unicommunity.repository.TemplateRepository
+import pt.isel.g20.unicommunity.template.exception.CannotModifyTemplatesBeingUsedException
 import pt.isel.g20.unicommunity.template.exception.NotFoundTemplateException
 import pt.isel.g20.unicommunity.template.model.Template
 
@@ -19,8 +20,13 @@ class TemplateService(val templatesRepo: TemplateRepository) : ITemplateService 
         return templatesRepo.save(template)
     }
 
-    override fun editTemplate(templateId: Long, hasForum: Boolean?, blackboardNames: String?): Template {
+    override fun editTemplate(templateId: Long, name: String, hasForum: Boolean?, blackboardNames: String?): Template {
         val template = getTemplateById(templateId)
+
+        template.name = name
+
+        if(template.boards.size != 0)
+            throw CannotModifyTemplatesBeingUsedException()
 
         if(hasForum != null)
             template.hasForum = hasForum
@@ -33,6 +39,9 @@ class TemplateService(val templatesRepo: TemplateRepository) : ITemplateService 
 
     override fun deleteTemplate(templateId: Long): Template {
         val template = getTemplateById(templateId)
+        if(template.boards.size != 0)
+            throw CannotModifyTemplatesBeingUsedException()
+
         templatesRepo.delete(template)
         return template
     }
